requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=1.3.0=pypi_0
addict=2.4.0=pypi_0
albucore=0.0.23=pypi_0
albumentations=2.0.3=pypi_0
annotated-types=0.7.0=pypi_0
attrs=25.1.0=pypi_0
basicsr=1.4.2=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.1.31=hbcca054_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.1=pypi_0
controlnet-aux=0.0.9=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
diffusers=0.32.2=pypi_0
easydict=1.13=pypi_0
einops=0.8.0=pypi_0
facexlib=0.3.0=pypi_0
filelock=3.17.0=pypi_0
filterpy=1.4.5=pypi_0
flatbuffers=25.1.24=pypi_0
fonttools=4.56.0=pypi_0
fsspec=2025.2.0=pypi_0
future=1.0.0=pypi_0
gfpgan=1.3.8=pypi_0
grpcio=1.70.0=pypi_0
huggingface-hub=0.28.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.6.1=pypi_0
insightface=0.7.3=pypi_0
ip-adapter=0.1.0=pypi_0
jax=0.5.0=pypi_0
jaxlib=0.5.0=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.43=h712a8e2_2
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
liblzma=5.6.4=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libsqlite=3.48.0=hee588c1_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
llvmlite=0.44.0=pypi_0
lmdb=1.6.2=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=pypi_0
mediapipe=0.10.21=pypi_0
ml-dtypes=0.5.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pypi_0
numba=0.61.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
onnx=1.17.0=pypi_0
onnxruntime=1.20.1=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.4.0=h7b32b05_1
opt-einsum=3.4.0=pypi_0
packaging=24.2=pypi_0
peft=0.14.0=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=pyh8b19718_0
platformdirs=4.3.6=pypi_0
prettytable=3.14.0=pypi_0
protobuf=4.25.6=pypi_0
psutil=6.1.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pyparsing=3.2.1=pypi_0
python=3.10.16=he725a3c_1_cpython
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h8228510_1
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.2=pypi_0
scikit-image=0.25.1=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.1=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.8.0=pyhff2d567_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
sounddevice=0.5.1=pypi_0
stringzilla=3.11.3=pypi_0
sympy=1.13.1=pypi_0
tb-nightly=2.19.0a20250207=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.1.10=pypi_0
timm=0.6.7=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.1.0=pypi_0
torchvision=0.16.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.48.3=pypi_0
trimesh=4.6.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025a=h78e105d_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
yapf=0.43.0=pypi_0
zipp=3.21.0=pypi_0
